Update IBIS and remove all Authentication components to leave an Access Control System only.
