+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "app/lib/advanced-custom-fields"]
 	path = app/lib/advanced-custom-fields
-	url = https://github.com/elliotcondon/acf.git
+	url = https://github.com/AdvancedCustomFields/acf.git
