--- app/error/auth_exception.php
@@ -1,21 +1,0 @@
-<?php
-/*
-    AuthException
-    -------------
-
-    Handles authentication errors.
-
-    @file    auth_exception.php
-    @license BSD 3-Clause
-    @package WPRavenAuth
-    @author  Gideon Farrell <me@gideonfarrell.co.uk>
- */
-
-namespace WPRavenAuth;
-    
-require('exception.php');
-
-class AuthException extends WPRException {
-    // TODO!
-}
-?>
